+++ config.env.template.js
@@ -1,5 +1,5 @@
 export const EnvConfig = {
+    baseUrl: "",
     mapApiKey: "",
     reCaptchaApiKey: "",
-    baseUrl: "",
 };
