+++ config/config.test.js
@@ -3,7 +3,7 @@ const CommonConfig = require("./config.common");
 const config = {
     ...CommonConfig,
     ...{
-        baseUrl: "http://localhost:3000"
+        baseUrl: "http://localhost:3000",
     },
 };
