+++ package.json
@@ -57,7 +57,7 @@
     "selenium-standalone": "^6.6.0",
     "sinon": "^3.2.0",
     "ts-node": "^3.3.0",
-    "tslint": "^5.7.0",
+    "tslint": "5.7.0",
     "tslint-react": "^3.2.0",
     "webpack-dev-server": "^2.7.1"
   },
