+++ src/data/Errors/SubmitError.ts
@@ -1,0 +1,6 @@
+export class SubmitError {
+    public constructor(sourceError: any, debugMode: boolean) {
+        // tslint:disable:no-console
+        debugMode && console.error(sourceError);
+    }
+}
